Move compiler class to own module
